+++ breadboard.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+#./start -Dapplication.secret="pleasechangethis" -Dconfig.file=application-prod.conf
+activator -Dhttps.port=9443 -Dapplication.secret="pleasechangethis" -Dconfig.file=application-prod.conf run